Update type & api dir structure
