+++ .eslintrc.js
@@ -6,7 +6,8 @@ module.exports = {
 		'react/require-default-props': 'off',
 		'import/no-unresolved': 'error',
 		'import/no-extraneous-dependencies': 'off',
-		'import/no-named-as-default': 'off'
+		'import/no-named-as-default': 'off',
+		'compat/compat': 'off'
 	},
 	parserOptions: {
 		ecmaVersion: 2020,
