+++ src/typings/metadata.ts
@@ -26,8 +26,7 @@ export interface AudioInfo {
 
 export interface AudioPicture {
 	format: string;
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	data: any;
+	data: string;
 }
 
 export interface AudioFormat {
