+++ src/ui/views/Home/Home.tsx
@@ -2,7 +2,7 @@ import { FC, useEffect, useState } from 'react';
 import { selectSpotifyToken, selectUser } from '../../../state/slices/playerSlice';
 import { IArtist } from '../../../typings/types';
 import { useAppSelector } from '../../hooks';
-import SpotifyAPI from '../../util/spotifyAPI';
+import SpotifyAPI from '../../api/spotify';
 
 import View from '../../components/View/View';
 import HomeItemMedium from '../../components/Home_Elements/HomeItemMedium/HomeItemMedium';
